server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Alex <alexander.ksenzov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
celery = "4.4.0"
django-admin-display = "^1.3.0"
strawberry-django-jwt = "^0.2.2"
PyJWT = "^2.5.0"
pytz = "^2022.2.1"
strawberry-graphql-django = "^0.5.2"
flower = "0.9.3"
redis = "^4.3.4"
django-cors-headers = "^3.13.0"
pytest = "^7.1.3"
Pillow = "^9.2.0"
django-allauth = "^0.51.0"
django-extensions = "^3.2.1"
django-extra-fields = "^3.0.2"
psycopg2 = "^2.9.3"
pytest-django = "^4.5.2"
strawberry-graphql = "^0.133.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pylint.MESSAGE_CONTROL]
disable = """
    line-too-long,
    abstract-method,
    no-member,
"""
enable = "useless-suppression"

[tool.pylint.REPORTS]
reports="no"

[tool.pylint.FORMAT]
max-line-length="88"

[tool.pylint.BASIC]
good-names = "_,i,e,setUp,tearDown,maxDiff"

[tool.black]
line-length = 88
include = '\.pyi?$'


[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
default_section = "THIRDPARTY"
known_first_party = []
known_third_party = ["PIL", "accounts", "celery", "django", "pytest", "strawberry", "strawberry_django_jwt"]
